Refactor WebSocket API
- Extract server into server.rs.
- Rename WebSocketAPI to WebSocketServer.
- Rename WebSocketConfig to ApiConfig.
- Add `struct ApiKey`.
- Rename `api_key` to `api_token` everywhere.
- Split load_api_key().
- Extract encode()/decode() functions.
- Change debug!() to trace!() for raw messages.
- Move NodeNotification and WalletNotification to ResponseKind in
  order to get the single structure for all outgoing messages.

Needed for remote console.
