+++ api/src/lib.rs
@@ -26,37 +26,22 @@
 mod config;
 mod crypto;
 mod error;
+mod server;
 
-use self::crypto::{decrypt, encrypt};
-pub use crate::config::WebSocketConfig;
+pub use crate::config::ApiConfig;
+pub use crate::crypto::ApiToken;
 pub use crate::error::KeyError;
-use failure::Error;
-use futures::sync::mpsc::UnboundedReceiver;
-use futures::sync::oneshot;
-use futures::{Async, AsyncSink, Future, Poll, Sink, Stream};
+pub use crate::server::WebSocketServer;
+
+use crate::crypto::{decrypt, encrypt};
 use log::*;
-use serde::Serialize;
+use serde::de::DeserializeOwned;
+use serde::ser::Serialize;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
-use serde_json;
-use std::net::SocketAddr;
-use stegos_node::{EpochChanged, Node, NodeRequest, NodeResponse, SyncChanged};
-use stegos_wallet::{Wallet, WalletNotification, WalletRequest, WalletResponse};
-use tokio::net::TcpListener;
-use tokio::runtime::TaskExecutor;
-use websocket::message::OwnedMessage;
-use websocket::result::WebSocketError;
-use websocket::server::upgrade::r#async::IntoWs;
-
-/// The number of values to fit in the output buffer.
-const OUTPUT_BUFFER_SIZE: usize = 10;
-/// Key size for API token
-const API_KEYSIZE: usize = 16;
-
-/// A type definition for sink.
-type WsSink = Box<Sink<SinkItem = OwnedMessage, SinkError = WebSocketError> + Send>;
-/// A type definition for stream.
-type WsStream = Box<Stream<Item = OwnedMessage, Error = WebSocketError> + Send>;
+use stegos_node::{EpochChanged, NodeRequest, NodeResponse, SyncChanged};
+use stegos_wallet::{WalletNotification, WalletRequest, WalletResponse};
+use websocket::WebSocketError;
 
 pub type RequestId = u64;
 
@@ -76,11 +61,21 @@ struct Request {
     id: u64,
 }
 
+#[derive(Debug, Clone, Serialize)]
+#[serde(tag = "notification")]
+#[serde(rename_all = "snake_case")]
+pub enum NodeNotification {
+    SyncChanged(SyncChanged),
+    EpochChanged(EpochChanged),
+}
+
 #[derive(Debug, Serialize)]
 #[serde(untagged)]
 enum ResponseKind {
     WalletResponse(WalletResponse),
     NodeResponse(NodeResponse),
+    NodeNotification(NodeNotification),
+    WalletNotification(WalletNotification),
 }
 
 fn is_default(id: &RequestId) -> bool {
@@ -96,317 +91,32 @@ struct Response {
     id: RequestId,
 }
 
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "notification")]
-#[serde(rename_all = "snake_case")]
-pub enum NodeNotification {
-    SyncChanged(SyncChanged),
-    EpochChanged(EpochChanged),
+fn encode<T: Serialize>(api_token: &ApiToken, msg: &T) -> String {
+    let msg = serde_json::to_vec(&msg).expect("serialized");
+    let msg = encrypt(api_token, &msg);
+    let msg = base64::encode(&msg);
+    msg
 }
 
-/// Handler of incoming connections.
-struct WebSocketHandler {
-    /// Remote address.
-    peer: SocketAddr,
-    /// API Key (32-bytes vector)
-    api_key: Vec<u8>,
-    /// Outgoing stream.
-    sink: WsSink,
-    /// Incoming stream.
-    stream: WsStream,
-    /// True if outgoing buffer should be flushed on the next poll().
-    need_flush: bool,
-    /// Wallet API.
-    wallet: Wallet,
-    /// Wallet events.
-    wallet_notifications: UnboundedReceiver<WalletNotification>,
-    /// Wallet RPC responses.
-    wallet_responses: Vec<(RequestId, oneshot::Receiver<WalletResponse>)>,
-    /// Node API.
-    node: Node,
-    /// Node RPC responses.
-    node_responses: Vec<(RequestId, oneshot::Receiver<NodeResponse>)>,
-    /// Synchronization Status Changed Notification.
-    node_sync_changed: UnboundedReceiver<SyncChanged>,
-    /// Epoch Changed Notification.
-    node_epoch_changed: UnboundedReceiver<EpochChanged>,
-}
-
-impl WebSocketHandler {
-    fn new(
-        peer: SocketAddr,
-        api_key: Vec<u8>,
-        sink: WsSink,
-        stream: WsStream,
-        wallet: Wallet,
-        node: Node,
-    ) -> Self {
-        let need_flush = false;
-        let wallet_notifications = wallet.subscribe();
-        let wallet_responses = Vec::new();
-        let node_responses = Vec::new();
-        let node_sync_changed = node.subscribe_sync_changed();
-        let node_epoch_changed = node.subscribe_epoch_changed();
-        WebSocketHandler {
-            peer,
-            api_key,
-            sink,
-            stream,
-            need_flush,
-            wallet,
-            wallet_notifications,
-            wallet_responses,
-            node,
-            node_responses,
-            node_sync_changed,
-            node_epoch_changed,
+fn decode<T: DeserializeOwned>(api_token: &ApiToken, msg: &str) -> Result<T, WebSocketError> {
+    let msg = match base64::decode(&msg) {
+        Ok(r) => r,
+        Err(e) => {
+            error!("Failed to base64::decode message: error={}", e);
+            return Err(WebSocketError::RequestError("Invalid request"));
         }
-    }
-
-    fn on_message(&mut self, enc_text: String) -> Result<(), WebSocketError> {
-        trace!("processing message: {}", enc_text);
-        let bin_text = match base64::decode(&enc_text) {
-            Ok(r) => r,
-            Err(e) => {
-                error!("Failed to base64::decode message: error={}", e);
-                return Err(WebSocketError::RequestError("Invalid request"));
-            }
-        };
-        let text = decrypt(&self.api_key, &bin_text);
-        let request: Request = match serde_json::from_slice(&text) {
-            Ok(r) => r,
-            Err(e) => {
-                error!(
-                    "Failed to deserialize: text={}, error={}",
-                    String::from_utf8_lossy(&text),
-                    e
-                );
-                return Err(WebSocketError::RequestError("Invalid request"));
-            }
-        };
-        match request.kind {
-            RequestKind::WalletRequest(wallet_request) => {
-                self.wallet_responses
-                    .push((request.id, self.wallet.request(wallet_request)));
-            }
-            RequestKind::NodeRequest(node_request) => {
-                self.node_responses
-                    .push((request.id, self.node.request(node_request)));
-            }
+    };
+    let msg = decrypt(api_token, &msg);
+    let msg: T = match serde_json::from_slice(&msg) {
+        Ok(r) => r,
+        Err(e) => {
+            error!(
+                "Failed to deserialize: msg={}, error={}",
+                String::from_utf8_lossy(&msg),
+                e
+            );
+            return Err(WebSocketError::RequestError("Invalid request"));
         }
-        Ok(())
-    }
-
-    fn send<T: Serialize>(&mut self, msg: T) {
-        trace!(
-            "sending message: {}",
-            serde_json::to_string(&msg).expect("serialized")
-        );
-        let msg = serde_json::to_vec(&msg).expect("serialized");
-        let enc_msg = encrypt(&self.api_key, &msg);
-        if let Err(e) = self.send_raw(OwnedMessage::Text(base64::encode(&enc_msg))) {
-            error!("Failed to send message: {}", e);
-        }
-    }
-
-    fn send_raw(&mut self, msg: OwnedMessage) -> Result<(), WebSocketError> {
-        trace!("raw send: {:?}", msg);
-        match self.sink.start_send(msg)? {
-            AsyncSink::Ready => {
-                self.need_flush = true;
-                Ok(())
-            }
-            AsyncSink::NotReady(msg) => {
-                warn!("The output buffer is full, discarding message: {:?}", msg);
-                Ok(())
-            }
-        }
-    }
-}
-
-impl Drop for WebSocketHandler {
-    fn drop(&mut self) {
-        info!("[{}] Disconnected", self.peer);
-    }
-}
-
-impl Future for WebSocketHandler {
-    type Item = ();
-    type Error = WebSocketError;
-
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        // Process incoming messages.
-        loop {
-            match self.stream.poll()? {
-                Async::Ready(Some(OwnedMessage::Text(msg))) => {
-                    debug!("[{}] Received text: {}", self.peer, &msg);
-                    self.on_message(msg)?
-                }
-                Async::Ready(Some(OwnedMessage::Binary(msg))) => {
-                    debug!("[{}] Received binary: len={}", self.peer, msg.len());
-                    return Err(WebSocketError::RequestError("BinaryIsNotSupported"));
-                }
-                Async::Ready(Some(OwnedMessage::Ping(msg))) => {
-                    debug!("[{}] Received ping: len={}", self.peer, msg.len());
-                    self.send_raw(OwnedMessage::Pong(msg))?
-                }
-                Async::Ready(Some(OwnedMessage::Pong(msg))) => {
-                    debug!("[{}] Received pong: len={}", self.peer, msg.len());
-                }
-                Async::Ready(Some(OwnedMessage::Close(data))) => {
-                    debug!(
-                        "[{}] Received close: has_data={}",
-                        self.peer,
-                        data.is_some()
-                    );
-                    return Ok(Async::Ready(()));
-                }
-                Async::Ready(None) => {
-                    debug!("[{}] Received eof", self.peer);
-                    return Ok(Async::Ready(()));
-                }
-                Async::NotReady => break,
-            }
-        }
-
-        loop {
-            match self.wallet_notifications.poll() {
-                Ok(Async::Ready(Some(notification))) => {
-                    self.send(&notification);
-                }
-                Ok(Async::Ready(None)) => return Ok(Async::Ready(())),
-                Ok(Async::NotReady) => break, // fall through
-                Err(()) => panic!("Wallet failure"),
-            }
-        }
-
-        let wallet_responses = std::mem::replace(&mut self.wallet_responses, Vec::new());
-        for (id, mut rx) in wallet_responses {
-            match rx.poll() {
-                Ok(Async::Ready(response)) => {
-                    let response = Response {
-                        kind: ResponseKind::WalletResponse(response),
-                        id,
-                    };
-                    self.send(response);
-                }
-                Ok(Async::NotReady) => self.wallet_responses.push((id, rx)),
-                Err(_) => panic!("disconnected"),
-            }
-        }
-
-        let node_responses = std::mem::replace(&mut self.node_responses, Vec::new());
-        for (id, mut rx) in node_responses {
-            match rx.poll() {
-                Ok(Async::Ready(response)) => {
-                    let response = Response {
-                        kind: ResponseKind::NodeResponse(response),
-                        id,
-                    };
-                    self.send(response)
-                }
-                Ok(Async::NotReady) => self.node_responses.push((id, rx)),
-                Err(_) => panic!("disconnected"),
-            }
-        }
-
-        // Height changes.
-        loop {
-            match self.node_sync_changed.poll().expect("connected") {
-                Async::Ready(Some(msg)) => {
-                    let msg = NodeNotification::SyncChanged(msg);
-                    self.send(msg);
-                }
-                Async::Ready(None) => return Ok(Async::Ready(())),
-                Async::NotReady => break, // fall through
-            }
-        }
-
-        // Epoch changes.
-        loop {
-            match self.node_epoch_changed.poll().expect("connected") {
-                Async::Ready(Some(msg)) => {
-                    let msg = NodeNotification::EpochChanged(msg);
-                    self.send(msg);
-                }
-                Async::Ready(None) => return Ok(Async::Ready(())),
-                Async::NotReady => break, // fall through
-            }
-        }
-
-        // Flush output buffer.
-        if self.need_flush {
-            match self.sink.poll_complete()? {
-                Async::Ready(()) => self.need_flush = false,
-                Async::NotReady => {}
-            }
-        }
-
-        Ok(Async::NotReady)
-    }
-}
-
-pub struct WebSocketAPI {}
-
-impl WebSocketAPI {
-    pub fn spawn(
-        cfg: WebSocketConfig,
-        executor: TaskExecutor,
-        wallet: Wallet,
-        node: Node,
-    ) -> Result<(), Error> {
-        let executor2 = executor.clone();
-        let wallet2 = wallet.clone();
-        let node2 = node.clone();
-        let addr: SocketAddr = format!("{}:{}", cfg.bind_ip, cfg.bind_port).parse()?;
-        let key = config::load_key(&cfg)?;
-        info!("Starting WebSocket API on {}", &addr);
-        let server = TcpListener::bind(&addr)?
-            .incoming()
-            .map_err(|e| {
-                error!("Failed to accept: {:?}", e);
-            })
-            .for_each(move |s| {
-                let wallet3 = wallet2.clone();
-                let node3 = node2.clone();
-                let peer = s.peer_addr().expect("has peer address");
-                let key = key.clone();
-                debug!("[{}] accepted", peer);
-                let s = s
-                    .into_ws()
-                    .map_err(move |(_s, _req, _buf, e)| {
-                        error!("[{}] Failed to upgrade to websocket: {}", &peer, e);
-                    })
-                    .and_then(move |upgrade| {
-                        upgrade
-                            .accept()
-                            .map(|(s, _headers)| s)
-                            .and_then(move |s| {
-                                let (sink, stream) = s.split();
-                                let sink = sink.buffer(OUTPUT_BUFFER_SIZE);
-                                let sink: WsSink = Box::new(sink);
-                                let stream: WsStream = Box::new(stream);
-                                info!("[{}] Connected", peer);
-                                WebSocketHandler::new(
-                                    peer,
-                                    key,
-                                    sink,
-                                    stream,
-                                    wallet3.clone(),
-                                    node3.clone(),
-                                )
-                            })
-                            .map_err(move |e| {
-                                error!("[{}] Error: {}", &peer, e);
-                            })
-                    });
-                // Spawn the connection handler.
-                executor2.spawn(s);
-                Ok(())
-            });
-
-        // Spawn the server.
-        executor.spawn(server);
-        Ok(())
-    }
+    };
+    Ok(msg)
 }
